# Conflicts:
#	src/main/java/app/AppBuilder.java
#	src/main/java/data_access/CSVTransactionDAO.java
#	src/main/java/data_access/InMemoryTransactionDataAccessObject.java
#	src/main/java/entity/Transaction.java
#	src/main/java/interface_adapter/TransactionDataAccess.java
#	src/main/java/interface_adapter/upload_statement/UploadStatementPresenter.java
#	src/main/java/use_case/upload_statement/UploadStatementInputBoundary.java
#	src/main/java/use_case/upload_statement/UploadStatementInteractor.java
#	src/main/java/use_case/upload_statement/UploadStatementOutputBoundary.java
#	src/main/java/view/UploadStatementView.java

+++ src/main/java/use_case/spending_report/GenerateReportInteractor.java
@@ -29,7 +29,7 @@ public class GenerateReportInteractor implements GenerateReportInputBoundary {
         Map<String, Float> categoryTotals = new HashMap<>();
         for (Transaction t : transactions) {
             String category = t.getCategory();
-            Float amount = t.getAmount();
+            Float amount = (float) t.getAmount();
             
             if (categoryTotals.containsKey(category)) {
                 categoryTotals.put(category, categoryTotals.get(category) + amount);
